requirements.txt
pycromanager==0.18.3
arduino-python3===0.6
